Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,7 +3,7 @@
 
 
 int main() {
-    float area;
+    int area;
     scanf("%area",&area);
     printf("Area: %f\n",area);
